New verstile testcase.
The new testcase is a program which uses BFD and libopcodes to carry
on a disassembly along with lend.

In this way, any other executable can be used as a testcase out of the box.
The ldtest program disassembles each instruction with both libopcodes
and lend, compares the lengths of the disassembled instructions, and in
case they are different highlights the output and shows the original
disassembled mnemonic. It points out as well the number of bytes
expected by lend and the number of actual bytes returned by it.

The length produces by libopcodes is used to recover from such errors and
continue disassemblying the original program.

An example .s has been introduced in the repository, which can be compiled
using any assembles (as or nasm) and passed as input to ldtest.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>