3-byte 3A modrm table, fix usage of SIB byte in mode 64
The table for 3-byte 3A opcode which tells whether a modrm byte is used
has been added to the decompiler.

Additionally, there is an issue with the usage of the SIB byte in 64 bits
mode, which is related to the fact that the 16-byte ModR/M addressing
is not actually used.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>