imm64 and moffs64 checked only on 1-byte opcodes
A multibyte instruction could have the actual last byte of the opcode
which clashes with that of the movabs. In this case, the disassembler
would think that there is an imm64 or a moffs64 operand, while there
isn't.
A flag now keeps track of a found multi-byte opcode, and jumps the
REX-related check to see if we're hitting a movabs.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>